# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'todo list api'
  version: 1.0.0
paths:
  /todos/:
    post:
      tags:
        - todos
      summary: Todoを作成する
      description: Todoとなる文字列を受け取り、Todoを作成する
      operationId: 'Create Todo'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreateRequest'
      responses:
        201:
          description: 'Successful operation'
  '/todos/{id}':
    delete:
      tags:
        - todos
      summary: todoを削除する
      description: todoを削除する
      operationId: Title
      parameters:
        -
          name: id
          in: path
          description: 削除対象のtodoのID
          required: true
          schema:
            type: int
          examples:
            int:
              summary: 'an int value'
              value: '1'
      responses:
        200:
          description: 削除成功
  /todos:
    get:
      tags:
        - todos
      summary: todoリストを取得する
      description: todoリストを取得する
      operationId: b03fcdee95a50470c67ff89fb50c0729
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        500:
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Internal Server Error'
                type: object
components:
  schemas:
    TodoCreateRequest:
      title: TodoCreateRequest
      description: Todo作成リクエスト
      required:
        - name
      properties:
        name:
          description: Todo名
          type: string
          example: 'Buy milk'
      type: object
    Todo:
      title: Todo
      description: 'Todo model'
      required:
        - 'id, name'
      properties:
        id:
          description: ID
          type: integer
          format: int64
          example: 1
        name:
          description: 'Todo name'
          type: string
          example: 'Buy milk'
      type: object
tags:
  -
    name: todos
    description: todos